+++ src/useEffectDebugger.ts
@@ -1,6 +1,16 @@
 import { useRef, useEffect } from 'react';
 
+const CONSOLE_NAME = 'use-effect-debugger';
+
+const CONSOLE_OUTPUTS = {
+  LOG: 'log',
+  TABLE: 'table',
+} as const;
+
+type ConsoleOutputs = typeof CONSOLE_OUTPUTS;
+
 type UseEffectDebuggerDebugOptions = {
+  consoleOutput?: ConsoleOutputs[keyof ConsoleOutputs];
   consoleName?: string;
   depNames?: (string | null)[];
 };
@@ -32,40 +42,55 @@ const usePreviousDeps = (
 const useEffectDebugger = (
   effect: React.EffectCallback,
   deps: React.DependencyList,
-  debugOptions?: UseEffectDebuggerDebugOptions
+  {
+    consoleOutput = CONSOLE_OUTPUTS.LOG,
+    consoleName = CONSOLE_NAME,
+    depNames = [],
+  }: UseEffectDebuggerDebugOptions = {}
 ) => {
   const prevDeps = usePreviousDeps(deps, []);
 
-  const consoleName = debugOptions?.consoleName ?? 'use-effect-debugger';
-  const depNames = debugOptions?.depNames ?? [];
-
   // Go through each of the deps to check and collate all of the deps that have changed
-  const changedDeps = deps.reduce<Record<string | number, unknown>>(
-    (acc, dep, idx) => {
-      // Check to see if the current value of the dep is NOT equal to the previous value of the dep
-      if (dep !== prevDeps[idx]) {
-        // Try to get the debug key name, else we fallback on the index within the deps array
-        const keyName = depNames[idx] ?? idx;
+  const changedDeps = deps.reduce<
+    Record<string | number, { prev: unknown; cur: unknown }>
+  >((acc, dep, idx) => {
+    // Check to see if the current value of the dep is NOT equal to the previous value of the dep
+    if (dep !== prevDeps[idx]) {
+      // Try to get the debug key name, else we fallback on the index within the deps array
+      const keyName = depNames[idx] ?? idx;
 
-        // Append the previous and current value of the dep, within the key name, to the accumulator
-        return {
-          ...acc,
-          [keyName]: {
-            prev: prevDeps[idx],
-            cur: dep,
-          },
-        };
-      }
+      // Append the previous and current value of the dep, within the key name, to the accumulator
+      return {
+        ...acc,
+        [keyName]: {
+          prev: prevDeps[idx],
+          cur: dep,
+        },
+      };
+    }
 
-      // If we get here, the previous and current dep values were equal and nothing needs to be added to the accumulator
-      return acc;
-    },
-    {}
-  );
+    // If we get here, the previous and current dep values were equal and nothing needs to be added to the accumulator
+    return acc;
+  }, {});
 
   // Log any changed deps to the console
   if (Object.keys(changedDeps).length) {
-    console.log(consoleName, changedDeps);
+    // Check what `consoleOutput` type was requested
+    switch (consoleOutput) {
+      // If `log`, use `console.log`
+      case CONSOLE_OUTPUTS.LOG:
+        console.log(consoleName, changedDeps);
+        break;
+      // If `table`, use `console.table`
+      case CONSOLE_OUTPUTS.TABLE:
+        // We have to log the `consoleName` and table seperately here as `console.table` doesn't support the name
+        console.log(consoleName);
+        console.table(changedDeps);
+        break;
+      // As a fallback/default, use `console.log`
+      default:
+        console.log(consoleName, changedDeps);
+    }
   }
 
   // Invoke the standard `useEffect` hook with the `effect` callback and `deps` passed